requirements.txt
# requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.6.1
xgboost==1.7.5
pyyaml==6.0
